key-Manager/requirements.txt
fastapi
uvicorn[standard]
motor
cryptography
argon2-cffi
python-dotenv

